Better handling of unknown colours
In case the user doesn't configure the ANSI palette properly, the colours
are unknown to tco and they're cannot be used for matching. However, they
should still be accessible through the '@X' notation.

This functionality was broken due to improper handling of them. This
commit fixes that. As a result, it also resolves #3.